Remove all imports of deferred_operations
